cleanup and fix bugs for runc and docker factory
